requirements.txt
openai>=1.0.0
litellm>=1.0.0
loguru>=0.7.0
pyyaml>=6.0
pydantic>=2.0
numpy>=1.24
pandas>=2.0
tqdm>=4.66
tomli>=2.0; python_version < "3.11"
